becker/max-retry
Dearest Reviewer,

Issue #19 requested a max retry. This is my take on a solution. I allows
the runner to set a max retry via sql. It does not check for valid sql
and if there is an error it just looks like the jobs are not running.

I used the stress test to test the change and I did not notice a
difference

Changes:
New config
New setter for config
Pass config in to mq_poll
use least to find the smaller number and return it.

Becker